"ads1_week3_exam: 2/5 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-1 assignments/M18 (exam 3)/Week-3_Assignment-1/Solution.java
@@ -14,14 +14,14 @@ class Stock implements Comparable<Stock> {
     /**
      * change in stock value.
      */
-    private int change;
+    private double change;
     /**
      * Constructs the object.
      *
      * @param      stocknamee  The stocknamee
      * @param      changee     The changee
      */
-    Stock(final String stocknamee, final int changee) {
+    Stock(final String stocknamee, final double changee) {
         this.stockname = stocknamee;
         this.change = changee;
     }
@@ -38,7 +38,7 @@ class Stock implements Comparable<Stock> {
      *
      * @return     The change.
      */
-    public int getChange() {
+    public double getChange() {
         return this.change;
     }
     /**
@@ -98,8 +98,8 @@ public final class Solution {
     public static void main(final String[] args) {
         Scanner scan = new Scanner(System.in);
         int n = Integer.parseInt(scan.nextLine());
-            BinarySearchST<String, Integer> best = new  BinarySearchST<>();
-            BinarySearchST<String, Integer> worst = new BinarySearchST<>();
+            BinarySearchST<String, Double> best = new  BinarySearchST<>();
+            BinarySearchST<String, Double> worst = new BinarySearchST<>();
         for (int i = 0; i < SIX; i++) {
             int count = 0;
             MinPQ<Stock> min = new MinPQ<>();
@@ -107,7 +107,7 @@ public final class Solution {
             while (count < n) {
                 String[] tokens = scan.nextLine().split(",");
                 Stock stocks = new Stock(tokens[0],
-                                         Integer.parseInt(tokens[1]));
+                                         Double.parseDouble(tokens[1]));
                 min.insert(stocks);
                 max.insert(stocks);
                 count++;
